requirements.txt
pandas==1.3.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.1
matplotlib==3.7.2
seaborn==0.12.2
Jinja2==3.0.1
openpyxl==3.1.2
xlrd==2.0.1
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
SQLAlchemy==1.4.23
gunicorn==21.2.0
python-dotenv==0.19.0
plotly==5.3.1
dash==2.0.0
dash-bootstrap-components==0.13.1
websockets==10.0
aiofiles==0.7.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
psutil==5.9.5 